Add comprehensive tests for session ingestion and enhance error handling
- Introduced a new test suite for session ingestion covering various edge cases and error handling scenarios.
- Added tests for logging different error severities in `_log_processing_error`.
- Implemented serialization tests for `ProcessingError` to ensure all fields are handled correctly.
- Enhanced rule loading tests to cover missing files and invalid YAML scenarios.
- Added tests for event processing to validate handling of non-dict payloads and unknown event types.
- Included tests for batch processing of events to ensure correct batch sizes and handling of empty iterators.
- Improved tests for file row creation and summary updates to validate correct behavior in various scenarios.
- Updated `test_codex_parser.py` to improve readability of bad event strings for invalid timestamp and payload types.
